fix: 보너스 볼 일치 여부를 확인하지 않는 로직 수정
