+++ src/main/java/application/Main.java
@@ -1,11 +1,6 @@
 package application;
 
 import application.routes.Route;
-import spark.ModelAndView;
-import spark.template.handlebars.HandlebarsTemplateEngine;
-
-import java.util.HashMap;
-import java.util.Map;
 
 import static spark.Spark.*;
